Byte Session: Routing Your Way Through Next JS and React Forms.
This was my second Byte Session on Next JS.

Tried to use `useLocalStorage` but I forgot that Next JS reconciles the front to the backend and you will get a mismatch error since the backend does not have local storage like the frontend. It's possible I could implement the local storage in a `useEffect` hook to handle client-side-only code, but trying to use a hook in a hook may not have been possible potentially. I believe there are packages that solve this problem or polyfill a local storage for the backend. I wanted to have a quick drop in for this by copying the hook from online, but it did not end up working out :RIP:.

I was hoping to have a users page that rendered all the users with their name, but after you click on them it could go to the dynamic page and display all of their information or even potentially allow you to edit it, but I ran into the persistent storage issue, and I think we ended on a good amount of time at 40 minutes; this would have pushed things out to an hour or more.

[changelog]
added:  Dependency: `formik`.
added: `Navigation` component.
added: `useLocalStorage` hook.
added: Input field example on the index page, showing use of `useState` with an initial value of `Coder`.
added: `Users` page with a form that matches up with the `User` interface.
added: Dynamic route `users/[id]` to demonstrate grabbing a dynamic segment of the URL utilizing `useRouter` via `router.query`.